requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.41.1
botocore==1.41.1
branca==0.8.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
entrypoints==0.4
Faker==38.2.0
favicon==0.7.0
filelock==3.20.0
folium==0.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2026.1.0
geopandas==1.1.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
postgrest==2.27.3
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.31.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymdown-extensions==10.17.1
pyogrio==0.11.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyproj==3.7.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.29.0
s3transfer==0.15.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.1.0
snowflake-snowpark-python==1.42.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
st-annotated-text==4.0.2
st-login-form==1.3.2
st-supabase-connection==2.1.3
st-theme==1.2.3
storage3==2.27.3
streamlit==1.51.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-folium==0.25.3
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-redirect==0.1.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.3
supabase-auth==2.27.3
supabase-functions==2.27.3
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
twilio==9.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
xyzservices==2025.11.0
yarl==1.22.0
